Code quality refactor
