+++ fitness_tracking/models/__init__.py
@@ -1,8 +1,10 @@
 """Fitness tracking SQLAlchemy 2.0 models."""
 
 # Import all models for SQLAlchemy discovery
-from .cricket import CricketCoachingEntry, CricketMatchEntry
-from .fitness import FitnessEntry, RestDayEntry
+from .cricket_coaching import CricketCoachingEntry
+from .cricket_match import CricketMatchEntry
+from .fitness import FitnessEntry
+from .rest_day import RestDayEntry
 
 __all__ = (
     "CricketCoachingEntry",
