+++ fitness_tracking/models/fitness.py
@@ -8,8 +8,8 @@ from sqlalchemy.orm import Mapped, mapped_column
 
 from common.models.activity import ActivityEntryBase
 from common.schemas.entry_type import EntryType
-from fitness_tracking.schemas.exercise_type import ExerciseType
-from fitness_tracking.schemas.intensity_level import IntensityLevel
+from fitness_tracking.schemas.enums.exercise_type import ExerciseType
+from fitness_tracking.schemas.enums.intensity_level import IntensityLevel
 
 
 class FitnessEntry(ActivityEntryBase):
@@ -58,44 +58,3 @@ class FitnessEntry(ActivityEntryBase):
     # Timing specifics
     start_time: Mapped[time | None] = mapped_column()
     end_time: Mapped[time | None] = mapped_column()
-
-
-class RestDayEntry(ActivityEntryBase):
-    """SQLAlchemy 2.0 model for rest day entries."""
-
-    __tablename__ = "rest_day_entries"
-
-    # Set the entry type
-    entry_type: Mapped[EntryType] = mapped_column(
-        SA_Enum(EntryType),
-        default=EntryType.REST_DAY,
-        index=True,
-    )
-
-    # Rest day specifics
-    rest_type: Mapped[str] = mapped_column(String)  # active, complete, partial
-    planned: Mapped[bool] = mapped_column(default=False)
-
-    # Recovery activities
-    recovery_activities: Mapped[list[str] | None] = mapped_column(JSON)
-    sleep_hours: Mapped[float | None] = mapped_column()
-    sleep_quality: Mapped[int | None] = mapped_column()  # 1-10 scale
-
-    # Physical state
-    muscle_soreness: Mapped[int | None] = mapped_column()  # 1-10 scale
-    fatigue_level: Mapped[int | None] = mapped_column()  # 1-10 scale
-    stress_level: Mapped[int | None] = mapped_column()  # 1-10 scale
-
-    # Recovery metrics
-    recovery_score: Mapped[int | None] = mapped_column()  # 1-100 scale
-    readiness_for_next_workout: Mapped[int | None] = mapped_column()  # 1-10 scale
-
-    # Wellness activities
-    meditation_minutes: Mapped[int | None] = mapped_column()
-    stretching_minutes: Mapped[int | None] = mapped_column()
-    massage_minutes: Mapped[int | None] = mapped_column()
-
-    # Nutrition focus
-    hydration_liters: Mapped[float | None] = mapped_column()
-    protein_focus: Mapped[bool | None] = mapped_column()
-    nutrition_notes: Mapped[str | None] = mapped_column(String)
