+++ fitness_tracking/repositories/__init__.py
@@ -1,11 +1,11 @@
 """Fitness tracking repositories for data access."""
 
-from .activity_repository import (
-    CricketCoachingEntryRepository,
+from .cricket_coaching_repository import CricketCoachingEntryRepository
+from .cricket_match_repository import (
     CricketMatchEntryRepository,
-    FitnessEntryRepository,
-    RestDayEntryRepository,
 )
+from .fitness_repository import FitnessEntryRepository
+from .rest_day_repository import RestDayEntryRepository
 
 __all__ = (
     "CricketCoachingEntryRepository",
