--- fitness_tracking/repositories/cricket_repository.py
@@ -1,1334 +1,0 @@
-"""Cricket repositories with analytics capabilities."""
-
-import logging
-from datetime import UTC, datetime, timedelta
-from typing import Any
-
-from sqlalchemy import func, select
-from sqlalchemy.ext.asyncio import AsyncSession
-
-from common.exceptions import AppError
-from common.repositories.crud_repository import CRUDRepository
-from fitness_tracking.models.cricket import (
-    CricketCoachingEntry,
-    CricketMatchEntry,
-)
-from fitness_tracking.models.fitness import RestDayEntry
-from fitness_tracking.schemas.cricket import (
-    CricketAnalytics,
-    CricketCoachingDataExtraction,
-    CricketCoachingEntryCreate,
-    CricketCoachingEntryRead,
-    CricketCoachingEntryUpdate,
-    CricketMatchDataExtraction,
-    CricketMatchEntryCreate,
-    CricketMatchEntryRead,
-    CricketMatchEntryUpdate,
-    RestDayDataExtraction,
-    RestDayEntryCreate,
-    RestDayEntryRead,
-    RestDayEntryUpdate,
-)
-from fitness_tracking.schemas.cricket_enums import (
-    CricketSessionType,
-    MatchFormat,
-    RestType,
-)
-
-logger = logging.getLogger(__name__)
-
-
-class CricketRepositoryError(AppError):
-    """Cricket repository specific error."""
-
-    status_code = 500
-    detail = "Cricket repository operation failed"
-
-
-class CricketCoachingNotFoundError(AppError):
-    """Cricket coaching entry not found error."""
-
-    status_code = 404
-    detail = "Cricket coaching entry not found"
-
-
-class CricketCoachingCreationError(AppError):
-    """Cricket coaching entry creation error."""
-
-    status_code = 400
-    detail = "Failed to create cricket coaching entry"
-
-
-class CricketMatchNotFoundError(AppError):
-    """Cricket match entry not found error."""
-
-    status_code = 404
-    detail = "Cricket match entry not found"
-
-
-class CricketMatchCreationError(AppError):
-    """Cricket match entry creation error."""
-
-    status_code = 400
-    detail = "Failed to create cricket match entry"
-
-
-class RestDayNotFoundError(AppError):
-    """Rest day entry not found error."""
-
-    status_code = 404
-    detail = "Rest day entry not found"
-
-
-class RestDayCreationError(AppError):
-    """Rest day entry creation error."""
-
-    status_code = 400
-    detail = "Failed to create rest day entry"
-
-
-class CricketCoachingRepository(
-    CRUDRepository[
-        CricketCoachingEntry,
-        CricketCoachingEntryCreate,
-        CricketCoachingEntryRead,
-        CricketCoachingEntryUpdate,
-    ],
-):
-    """Repository for cricket coaching session operations."""
-
-    def __init__(self, session: AsyncSession) -> None:
-        """Initialize cricket coaching repository."""
-        super().__init__(
-            model=CricketCoachingEntry,
-            create_schema=CricketCoachingEntryCreate,
-            read_schema=CricketCoachingEntryRead,
-            update_schema=CricketCoachingEntryUpdate,
-            session=session,
-            not_found_exception=CricketCoachingNotFoundError,
-            creation_exception=CricketCoachingCreationError,
-            user_filter=self._user_filter,
-        )
-
-    def _user_filter(self, user: Any) -> Any:
-        """Filter by user."""
-        if user:
-            return CricketCoachingEntry.user_id == user.id
-        return True
-
-    def _normalize_session_type(self, session_type: str) -> str:
-        """Normalize session type strings to valid values."""
-        session_type_lower = session_type.lower().strip()
-
-        # Mapping of common variations to valid enum values
-        session_type_mapping = {
-            # Batting variations
-            "batting": CricketSessionType.BATTING_DRILLS.value,
-            "batting_drills": CricketSessionType.BATTING_DRILLS.value,
-            "batting drills": CricketSessionType.BATTING_DRILLS.value,
-            "batting practice": CricketSessionType.BATTING_DRILLS.value,
-            "bat": CricketSessionType.BATTING_DRILLS.value,
-            "stroke play": CricketSessionType.BATTING_DRILLS.value,
-            "technique": CricketSessionType.BATTING_DRILLS.value,
-            # Wicket keeping variations
-            "keeping": CricketSessionType.WICKET_KEEPING.value,
-            "wicket_keeping": CricketSessionType.WICKET_KEEPING.value,
-            "wicket keeping": CricketSessionType.WICKET_KEEPING.value,
-            "wicketkeeping": CricketSessionType.WICKET_KEEPING.value,
-            "gloves": CricketSessionType.WICKET_KEEPING.value,
-            "keeper": CricketSessionType.WICKET_KEEPING.value,
-            # Netting variations
-            "nets": CricketSessionType.NETTING.value,
-            "netting": CricketSessionType.NETTING.value,
-            "net session": CricketSessionType.NETTING.value,
-            "net practice": CricketSessionType.NETTING.value,
-            # Personal coaching variations
-            "personal": CricketSessionType.PERSONAL_COACHING.value,
-            "personal_coaching": CricketSessionType.PERSONAL_COACHING.value,
-            "personal coaching": CricketSessionType.PERSONAL_COACHING.value,
-            "1-on-1": CricketSessionType.PERSONAL_COACHING.value,
-            "one on one": CricketSessionType.PERSONAL_COACHING.value,
-            "individual": CricketSessionType.PERSONAL_COACHING.value,
-            "private": CricketSessionType.PERSONAL_COACHING.value,
-            # Team practice variations
-            "team": CricketSessionType.TEAM_PRACTICE.value,
-            "team_practice": CricketSessionType.TEAM_PRACTICE.value,
-            "team practice": CricketSessionType.TEAM_PRACTICE.value,
-            "squad": CricketSessionType.TEAM_PRACTICE.value,
-            "group": CricketSessionType.TEAM_PRACTICE.value,
-            "training": CricketSessionType.TEAM_PRACTICE.value,
-            # Other/general variations
-            "other": CricketSessionType.OTHER.value,
-            "general": CricketSessionType.OTHER.value,
-            "mixed": CricketSessionType.OTHER.value,
-            "various": CricketSessionType.OTHER.value,
-        }
-
-        normalized = session_type_mapping.get(session_type_lower)
-        if normalized:
-            return normalized
-
-        # Try partial matching for compound terms
-        if any(keyword in session_type_lower for keyword in ["batting", "bat", "stroke"]):
-            return CricketSessionType.BATTING_DRILLS.value
-        if any(keyword in session_type_lower for keyword in ["keeping", "keeper", "glove"]):
-            return CricketSessionType.WICKET_KEEPING.value
-        if any(keyword in session_type_lower for keyword in ["net", "practice"]):
-            return CricketSessionType.NETTING.value
-        if any(keyword in session_type_lower for keyword in ["team", "squad", "group"]):
-            return CricketSessionType.TEAM_PRACTICE.value
-        if any(keyword in session_type_lower for keyword in ["personal", "individual", "private"]):
-            return CricketSessionType.PERSONAL_COACHING.value
-
-        # Default fallback
-        logger.warning("Unknown session type '%s', defaulting to 'other'", session_type)
-        return CricketSessionType.OTHER.value
-
-    def _normalize_integer_field(
-        self,
-        value: str | int | None,
-        field_name: str,
-        min_val: int = 1,
-        max_val: int = 10,
-        default: int = 5,
-    ) -> int:
-        """Normalize integer fields that might come as strings from LLM."""
-        logger.info(f"🔧 Normalizing {field_name}: {value} (type: {type(value)})")
-
-        if value is None:
-            logger.info(f"🔧 {field_name} is None, returning default: {default}")
-            return default
-
-        if isinstance(value, int):
-            result = max(min_val, min(max_val, value))
-            logger.info(f"🔧 {field_name} is int, returning clamped: {result}")
-            return result
-
-        if isinstance(value, str):
-            value_lower = value.lower().strip()
-            logger.info(f"🔧 {field_name} is string, processing: '{value_lower}'")
-
-            # Try to extract number from string
-            import re
-
-            number_match = re.search(r"\d+", value_lower)
-            if number_match:
-                try:
-                    num = int(number_match.group())
-                    result = max(min_val, min(max_val, num))
-                    logger.info(f"🔧 {field_name} extracted number: {result}")
-                    return result
-                except ValueError:
-                    pass
-
-            # Map descriptive terms to numbers - coaching specific
-            rating_mapping = {
-                "terrible": 1,
-                "awful": 1,
-                "very poor": 1,
-                "poor": 2,
-                "bad": 2,
-                "low": 2,
-                "below average": 3,
-                "not good": 3,
-                "weak": 3,
-                "average": 5,
-                "medium": 5,
-                "okay": 5,
-                "decent": 5,
-                "fine": 5,
-                "good": 6,
-                "well": 6,
-                "nice": 6,
-                "very good": 7,
-                "great": 7,
-                "strong": 7,
-                "excellent": 8,
-                "outstanding": 8,
-                "fantastic": 8,
-                "amazing": 9,
-                "incredible": 9,
-                "superb": 9,
-                "perfect": 10,
-                "flawless": 10,
-                "best": 10,
-            }
-
-            for term, rating in rating_mapping.items():
-                if term in value_lower:
-                    return max(min_val, min(max_val, rating))
-
-        logger.warning(
-            "Could not normalize %s value '%s', using default %d",
-            field_name,
-            value,
-            default,
-        )
-        return default
-
-    async def create_from_voice_data(
-        self,
-        session_id: str,
-        user_id: str,
-        voice_data: CricketCoachingDataExtraction | dict[str, Any],
-        transcript: str,
-        confidence_score: float,
-        processing_duration: float,
-    ) -> CricketCoachingEntry:
-        """Create cricket coaching entry from voice processing data."""
-        try:
-            # Handle both Pydantic model and dict input
-            if isinstance(voice_data, CricketCoachingDataExtraction):
-                data = voice_data.model_dump()
-            else:
-                data = voice_data
-
-            # Normalize session type
-            session_type = self._normalize_session_type(data.get("session_type", "team_practice"))
-
-            # Normalize integer fields that might come as strings from LLM
-            duration_minutes = self._normalize_integer_field(
-                data.get("duration_minutes"),
-                "duration_minutes",
-                5,
-                300,
-                60,
-            )
-            self_assessment_score = self._normalize_integer_field(
-                data.get("self_assessment_score"),
-                "self_assessment_score",
-                1,
-                10,
-                6,
-            )
-            confidence_level = self._normalize_integer_field(
-                data.get("confidence_level"),
-                "confidence_level",
-                1,
-                10,
-                6,
-            )
-            focus_level = self._normalize_integer_field(
-                data.get("focus_level"),
-                "focus_level",
-                1,
-                10,
-                6,
-            )
-            learning_satisfaction = self._normalize_integer_field(
-                data.get("learning_satisfaction"),
-                "learning_satisfaction",
-                1,
-                10,
-                6,
-            )
-            difficulty_level = self._normalize_integer_field(
-                data.get("difficulty_level"),
-                "difficulty_level",
-                1,
-                10,
-                5,
-            )
-
-            # Create cricket coaching entry
-            coaching_entry = CricketCoachingEntry(
-                session_id=session_id,
-                user_id=user_id,
-                # Voice processing metadata
-                transcript=transcript,
-                confidence_score=confidence_score,
-                processing_duration=processing_duration,
-                # Cricket coaching data
-                session_type=session_type,
-                duration_minutes=duration_minutes,
-                what_went_well=data.get("what_went_well", ""),
-                areas_for_improvement=data.get("areas_for_improvement", ""),
-                coach_feedback=data.get("coach_feedback"),
-                self_assessment_score=self_assessment_score,
-                skills_practiced=data.get("skills_practiced", ""),
-                difficulty_level=difficulty_level,
-                confidence_level=confidence_level,
-                focus_level=focus_level,
-                learning_satisfaction=learning_satisfaction,
-                mental_state=data.get("mental_state", "good"),
-                notes=data.get("notes"),
-            )
-
-            self.session.add(coaching_entry)
-            await self.session.commit()
-            await self.session.refresh(coaching_entry)
-
-            logger.info("Created cricket coaching entry from voice data: %s", coaching_entry.id)
-            return coaching_entry
-
-        except Exception as e:
-            await self.session.rollback()
-            logger.exception("Failed to create cricket coaching entry from voice data")
-            msg = f"Failed to create cricket coaching entry: {e}"
-            raise CricketRepositoryError(msg) from e
-
-    # The _to_read_schema method is inherited from CRUDRepository
-
-
-class CricketMatchRepository(
-    CRUDRepository[
-        CricketMatchEntry,
-        CricketMatchEntryCreate,
-        CricketMatchEntryRead,
-        CricketMatchEntryUpdate,
-    ],
-):
-    """Repository for cricket match performance operations."""
-
-    def __init__(self, session: AsyncSession) -> None:
-        """Initialize cricket match repository."""
-        super().__init__(
-            model=CricketMatchEntry,
-            create_schema=CricketMatchEntryCreate,
-            read_schema=CricketMatchEntryRead,
-            update_schema=CricketMatchEntryUpdate,
-            session=session,
-            not_found_exception=CricketMatchNotFoundError,
-            creation_exception=CricketMatchCreationError,
-            user_filter=self._user_filter,
-        )
-
-    def _user_filter(self, user: Any) -> Any:
-        """Filter by user."""
-        if user:
-            return CricketMatchEntry.user_id == user.id
-        return True
-
-    def _normalize_match_type(self, match_type: str) -> MatchFormat:
-        """Normalize match type strings to valid MatchFormat enum values."""
-        match_type_lower = match_type.lower().strip()
-
-        # Mapping of common variations to valid enum values
-        match_type_mapping = {
-            # Practice variations
-            "practice": MatchFormat.PRACTICE_MATCH,
-            "practice_match": MatchFormat.PRACTICE_MATCH,
-            "practice match": MatchFormat.PRACTICE_MATCH,
-            "friendly": MatchFormat.FRIENDLY,
-            "casual": MatchFormat.FRIENDLY,
-            "informal": MatchFormat.FRIENDLY,
-            "nets": MatchFormat.PRACTICE_MATCH,
-            # T20 variations
-            "t20": MatchFormat.T20,
-            "twenty20": MatchFormat.T20,
-            "20-20": MatchFormat.T20,
-            "20/20": MatchFormat.T20,
-            # ODI variations
-            "odi": MatchFormat.ODI,
-            "one day": MatchFormat.ODI,
-            "one-day": MatchFormat.ODI,
-            "50 over": MatchFormat.ODI,
-            "50-over": MatchFormat.ODI,
-            # Test variations
-            "test": MatchFormat.TEST,
-            "test match": MatchFormat.TEST,
-            "5 day": MatchFormat.TEST,
-            "five day": MatchFormat.TEST,
-            # List A variations
-            "list a": MatchFormat.LIST_A,
-            "list-a": MatchFormat.LIST_A,
-            "lista": MatchFormat.LIST_A,
-            # First class variations
-            "first class": MatchFormat.FIRST_CLASS,
-            "first-class": MatchFormat.FIRST_CLASS,
-            "fc": MatchFormat.FIRST_CLASS,
-            "firstclass": MatchFormat.FIRST_CLASS,
-            # Tournament
-            "tournament": MatchFormat.T20,  # Default tournament to T20
-            "competition": MatchFormat.T20,
-            "league": MatchFormat.T20,
-            "championship": MatchFormat.T20,
-            # School/College defaults to practice match
-            "school": MatchFormat.PRACTICE_MATCH,
-            "college": MatchFormat.PRACTICE_MATCH,
-            "university": MatchFormat.PRACTICE_MATCH,
-            "inter-school": MatchFormat.PRACTICE_MATCH,
-            "academic": MatchFormat.PRACTICE_MATCH,
-            # Club matches
-            "club": MatchFormat.FRIENDLY,
-            "local": MatchFormat.FRIENDLY,
-            "community": MatchFormat.FRIENDLY,
-            "district": MatchFormat.FRIENDLY,
-        }
-
-        normalized = match_type_mapping.get(match_type_lower)
-        if normalized:
-            return normalized
-
-        # Try partial matching for compound terms
-        if any(
-            keyword in match_type_lower
-            for keyword in ["friendly", "practice", "casual", "informal", "nets"]
-        ):
-            return MatchFormat.PRACTICE_MATCH
-        if any(keyword in match_type_lower for keyword in ["t20", "twenty", "20"]):
-            return MatchFormat.T20
-        if any(keyword in match_type_lower for keyword in ["odi", "one day", "50"]):
-            return MatchFormat.ODI
-        if any(keyword in match_type_lower for keyword in ["test", "5 day", "five day"]):
-            return MatchFormat.TEST
-        if any(keyword in match_type_lower for keyword in ["first class", "first-class"]):
-            return MatchFormat.FIRST_CLASS
-        if any(
-            keyword in match_type_lower
-            for keyword in ["tournament", "competition", "league", "championship"]
-        ):
-            return MatchFormat.T20  # Default tournament to T20
-        if any(
-            keyword in match_type_lower
-            for keyword in ["school", "college", "university", "academic"]
-        ):
-            return MatchFormat.PRACTICE_MATCH
-        if any(
-            keyword in match_type_lower for keyword in ["club", "local", "community", "district"]
-        ):
-            return MatchFormat.FRIENDLY
-
-        # Default fallback
-        logger.warning("Unknown match type '%s', defaulting to 'friendly'", match_type)
-        return MatchFormat.FRIENDLY
-
-    def _normalize_integer_field(
-        self,
-        value: str | int | None,
-        field_name: str,
-        min_val: int = 1,
-        max_val: int = 10,
-        default: int = 5,
-    ) -> int:
-        """Normalize integer fields that might come as strings from LLM."""
-        logger.info(f"🔧 Normalizing {field_name}: {value} (type: {type(value)})")
-
-        if value is None:
-            logger.info(f"🔧 {field_name} is None, returning default: {default}")
-            return default
-
-        if isinstance(value, int):
-            result = max(min_val, min(max_val, value))
-            logger.info(f"🔧 {field_name} is int, returning clamped: {result}")
-            return result
-
-        if isinstance(value, str):
-            value_lower = value.lower().strip()
-            logger.info(f"🔧 {field_name} is string, processing: '{value_lower}'")
-
-            # Try to extract number from string
-            import re
-
-            number_match = re.search(r"\d+", value_lower)
-            if number_match:
-                try:
-                    num = int(number_match.group())
-                    result = max(min_val, min(max_val, num))
-                    logger.info(f"🔧 {field_name} extracted number: {result}")
-                    return result
-                except ValueError:
-                    pass
-
-            # Map descriptive terms to numbers based on field type
-            if field_name == "opposition_strength":
-                strength_mapping = {
-                    "very weak": 1,
-                    "weak": 2,
-                    "poor": 2,
-                    "not very strong": 3,
-                    "below average": 3,
-                    "easy": 3,
-                    "average": 5,
-                    "medium": 5,
-                    "okay": 5,
-                    "decent": 5,
-                    "good": 6,
-                    "strong": 7,
-                    "tough": 7,
-                    "very strong": 8,
-                    "excellent": 9,
-                    "outstanding": 10,
-                    "best": 10,
-                    "top": 10,
-                }
-                logger.info(f"🔧 {field_name} checking against strength mapping")
-                for term, rating in strength_mapping.items():
-                    if term in value_lower:
-                        logger.info(f"🔧 {field_name} matched '{term}' -> {rating}")
-                        return rating
-                logger.info(f"🔧 {field_name} no mapping found for '{value_lower}'")
-
-            elif field_name in ["pre_match_nerves", "post_match_satisfaction"]:
-                emotion_mapping = {
-                    "none": 1,
-                    "not at all": 1,
-                    "zero": 1,
-                    "very low": 2,
-                    "low": 3,
-                    "little": 3,
-                    "some": 4,
-                    "medium": 5,
-                    "average": 5,
-                    "okay": 5,
-                    "high": 7,
-                    "very high": 8,
-                    "extremely": 9,
-                    "maximum": 10,
-                    "not very satisfied": 3,
-                    "not satisfied": 2,
-                    "disappointed": 2,
-                    "satisfied": 6,
-                    "very satisfied": 8,
-                    "extremely satisfied": 9,
-                    "confident": 7,
-                    "very confident": 8,
-                    "nervous": 3,
-                    "very nervous": 2,
-                }
-                logger.info(f"🔧 {field_name} checking against emotion mapping")
-                for term, rating in emotion_mapping.items():
-                    if term in value_lower:
-                        logger.info(f"🔧 {field_name} matched '{term}' -> {rating}")
-                        return rating
-                logger.info(f"🔧 {field_name} no emotion mapping found for '{value_lower}'")
-
-        logger.warning(
-            "Could not normalize %s value '%s', using default %d",
-            field_name,
-            value,
-            default,
-        )
-        return default
-
-    async def create_from_voice_data(
-        self,
-        session_id: str,
-        user_id: str,
-        voice_data: CricketMatchDataExtraction | dict[str, Any],
-        transcript: str,
-        confidence_score: float,
-        processing_duration: float,
-    ) -> CricketMatchEntry:
-        """Create cricket match entry from voice processing data."""
-        try:
-            logger.info(f"🏏 Received voice_data type: {type(voice_data)}")
-            logger.info(f"🏏 Raw voice_data: {voice_data}")
-
-            # Handle both Pydantic model and dict input
-            if isinstance(voice_data, CricketMatchDataExtraction):
-                data = voice_data.model_dump()
-                logger.info(f"🏏 Converted Pydantic model to dict: {data}")
-            else:
-                data = voice_data
-                logger.info(f"🏏 Using dict directly: {data}")
-
-            # Normalize match type
-            raw_match_type = data.get("match_type", "school")
-            logger.info(f"🏏 Raw match_type: {raw_match_type} (type: {type(raw_match_type)})")
-            match_type = self._normalize_match_type(raw_match_type)
-            logger.info(f"🏏 Normalized match_type: {match_type}")
-
-            # Normalize integer fields that might come as strings from LLM
-            raw_opposition_strength = data.get("opposition_strength")
-            logger.info(
-                f"🏏 Raw opposition_strength: {raw_opposition_strength} (type: {type(raw_opposition_strength)})",
-            )
-            opposition_strength = self._normalize_integer_field(
-                raw_opposition_strength,
-                "opposition_strength",
-                1,
-                10,
-                5,
-            )
-            logger.info(f"🏏 Normalized opposition_strength: {opposition_strength}")
-
-            raw_pre_match_nerves = data.get("pre_match_nerves")
-            logger.info(
-                f"🏏 Raw pre_match_nerves: {raw_pre_match_nerves} (type: {type(raw_pre_match_nerves)})",
-            )
-            pre_match_nerves = self._normalize_integer_field(
-                raw_pre_match_nerves,
-                "pre_match_nerves",
-                1,
-                10,
-                5,
-            )
-            logger.info(f"🏏 Normalized pre_match_nerves: {pre_match_nerves}")
-
-            raw_post_match_satisfaction = data.get("post_match_satisfaction")
-            logger.info(
-                f"🏏 Raw post_match_satisfaction: {raw_post_match_satisfaction} (type: {type(raw_post_match_satisfaction)})",
-            )
-            post_match_satisfaction = self._normalize_integer_field(
-                raw_post_match_satisfaction,
-                "post_match_satisfaction",
-                1,
-                10,
-                5,
-            )
-            logger.info(f"🏏 Normalized post_match_satisfaction: {post_match_satisfaction}")
-
-            # Create cricket match entry
-            match_entry = CricketMatchEntry(
-                session_id=session_id,
-                user_id=user_id,
-                # Voice processing metadata
-                transcript=transcript,
-                confidence_score=confidence_score,
-                processing_duration=processing_duration,
-                # Cricket match data
-                match_type=match_type,
-                opposition_strength=opposition_strength,
-                pre_match_nerves=pre_match_nerves,
-                post_match_satisfaction=post_match_satisfaction,
-                mental_state=data.get("mental_state", "good"),
-                # Batting stats
-                runs_scored=data.get("runs_scored"),
-                balls_faced=data.get("balls_faced"),
-                boundaries_4s=data.get("boundaries_4s"),
-                boundaries_6s=data.get("boundaries_6s"),
-                how_out=data.get("how_out"),
-                key_shots_played=data.get("key_shots_played"),
-                # Wicket keeping stats
-                catches_taken=data.get("catches_taken"),
-                catches_dropped=data.get("catches_dropped"),
-                stumpings=data.get("stumpings"),
-                notes=data.get("notes"),
-            )
-
-            logger.info(
-                f"🏏 Final match_entry object created with opposition_strength: {match_entry.opposition_strength}",
-            )
-
-            self.session.add(match_entry)
-            await self.session.commit()
-            await self.session.refresh(match_entry)
-
-            logger.info("Created cricket match entry from voice data: %s", match_entry.id)
-            return match_entry
-
-        except Exception as e:
-            await self.session.rollback()
-            logger.exception("Failed to create cricket match entry from voice data")
-            msg = f"Failed to create cricket match entry: {e}"
-            raise CricketRepositoryError(msg) from e
-
-    def _to_read_schema(self, db_record: CricketMatchEntry) -> CricketMatchEntryRead:
-        """Convert database model to read schema."""
-        return CricketMatchEntryRead(
-            id=db_record.id,
-            session_id=db_record.session_id,
-            user_id=db_record.user_id,
-            match_type=db_record.match_type.value,
-            opposition_strength=db_record.opposition_strength,
-            pre_match_nerves=db_record.pre_match_nerves,
-            post_match_satisfaction=db_record.post_match_satisfaction,
-            mental_state=db_record.mental_state,
-            runs_scored=db_record.runs_scored,
-            balls_faced=db_record.balls_faced,
-            boundaries_4s=db_record.boundaries_4s,
-            boundaries_6s=db_record.boundaries_6s,
-            how_out=db_record.how_out,
-            key_shots_played=db_record.key_shots_played,
-            catches_taken=db_record.catches_taken,
-            catches_dropped=db_record.catches_dropped,
-            stumpings=db_record.stumpings,
-            notes=db_record.notes,
-            transcript=db_record.transcript,
-            confidence_score=db_record.confidence_score,
-            processing_duration=db_record.processing_duration,
-            created_at=db_record.created_at,
-            updated_at=db_record.updated_at,
-        )
-
-
-class RestDayRepository(
-    CRUDRepository[RestDayEntry, RestDayEntryCreate, RestDayEntryRead, RestDayEntryUpdate],
-):
-    """Repository for rest day operations."""
-
-    def __init__(self, session: AsyncSession) -> None:
-        """Initialize rest day repository."""
-        super().__init__(
-            model=RestDayEntry,
-            create_schema=RestDayEntryCreate,
-            read_schema=RestDayEntryRead,
-            update_schema=RestDayEntryUpdate,
-            session=session,
-            not_found_exception=RestDayNotFoundError,
-            creation_exception=RestDayCreationError,
-            user_filter=self._user_filter,
-        )
-
-    def _user_filter(self, user):
-        """Filter by user."""
-        if user:
-            return RestDayEntry.user_id == user.id
-        return True
-
-    def _normalize_rest_type(self, rest_type: str) -> RestType:
-        """Normalize rest type strings to valid RestType enum values."""
-        rest_type_lower = rest_type.lower().strip()
-
-        # Mapping of common variations to valid enum values
-        rest_type_mapping = {
-            # Complete rest variations
-            "complete_rest": RestType.COMPLETE.value,
-            "complete rest": RestType.COMPLETE.value,
-            "full rest": RestType.COMPLETE.value,
-            "total rest": RestType.COMPLETE.value,
-            "no activity": RestType.COMPLETE.value,
-            "rest": RestType.COMPLETE.value,
-            "off": RestType.COMPLETE.value,
-            "lazy": RestType.COMPLETE.value,
-            "chill": RestType.COMPLETE.value,
-            # Active recovery variations
-            "active_recovery": RestType.ACTIVE.value,
-            "active recovery": RestType.ACTIVE.value,
-            "light activity": RestType.ACTIVE.value,
-            "easy day": RestType.ACTIVE.value,
-            "gentle": RestType.ACTIVE.value,
-            "walk": RestType.ACTIVE.value,
-            "walking": RestType.ACTIVE.value,
-            "stretch": RestType.ACTIVE.value,
-            "stretching": RestType.ACTIVE.value,
-            "yoga": RestType.ACTIVE.value,
-            "mobility": RestType.ACTIVE.value,
-            # Injury recovery variations
-            "injury_recovery": RestType.RECOVERY_DAY.value,
-            "injury recovery": RestType.RECOVERY_DAY.value,
-            "hurt": RestType.RECOVERY_DAY.value,
-            "injured": RestType.RECOVERY_DAY.value,
-            "pain": RestType.RECOVERY_DAY.value,
-            "sore": RestType.RECOVERY_DAY.value,
-            "rehab": RestType.RECOVERY_DAY.value,
-            "rehabilitation": RestType.RECOVERY_DAY.value,
-            "healing": RestType.RECOVERY_DAY.value,
-            "recovery": RestType.RECOVERY_DAY.value,
-        }
-
-        normalized = rest_type_mapping.get(rest_type_lower)
-        if normalized:
-            return normalized
-
-        # Try partial matching for compound terms
-        if any(keyword in rest_type_lower for keyword in ["complete", "full", "total", "off"]):
-            return RestType.COMPLETE.value
-        if any(
-            keyword in rest_type_lower
-            for keyword in ["active", "light", "gentle", "walk", "stretch", "yoga"]
-        ):
-            return RestType.ACTIVE.value
-        if any(
-            keyword in rest_type_lower for keyword in ["injury", "hurt", "pain", "rehab", "healing"]
-        ):
-            return RestType.RECOVERY_DAY.value
-
-        # Default fallback
-        logger.warning("Unknown rest type '%s', defaulting to 'complete_rest'", rest_type)
-        return RestType.COMPLETE.value
-
-    def _normalize_integer_field(
-        self,
-        value: str | int | None,
-        field_name: str,
-        min_val: int = 1,
-        max_val: int = 10,
-        default: int = 5,
-    ) -> int:
-        """Normalize integer fields that might come as strings from LLM."""
-        logger.info(f"🔧 Normalizing {field_name}: {value} (type: {type(value)})")
-
-        if value is None:
-            logger.info(f"🔧 {field_name} is None, returning default: {default}")
-            return default
-
-        if isinstance(value, int):
-            result = max(min_val, min(max_val, value))
-            logger.info(f"🔧 {field_name} is int, returning clamped: {result}")
-            return result
-
-        if isinstance(value, str):
-            value_lower = value.lower().strip()
-            logger.info(f"🔧 {field_name} is string, processing: '{value_lower}'")
-
-            # Try to extract number from string
-            import re
-
-            number_match = re.search(r"\d+", value_lower)
-            if number_match:
-                try:
-                    num = int(number_match.group())
-                    result = max(min_val, min(max_val, num))
-                    logger.info(f"🔧 {field_name} extracted number: {result}")
-                    return result
-                except ValueError:
-                    pass
-
-            # Map descriptive terms to numbers - rest day specific
-            level_mapping = {
-                "none": 1,
-                "zero": 1,
-                "not at all": 1,
-                "very low": 2,
-                "minimal": 2,
-                "barely": 2,
-                "low": 3,
-                "little": 3,
-                "slight": 3,
-                "moderate": 4,
-                "some": 4,
-                "a bit": 4,
-                "medium": 5,
-                "average": 5,
-                "okay": 5,
-                "normal": 5,
-                "high": 6,
-                "quite": 6,
-                "fairly": 6,
-                "very high": 7,
-                "strong": 7,
-                "significant": 7,
-                "extremely high": 8,
-                "intense": 8,
-                "severe": 8,
-                "maximum": 9,
-                "overwhelming": 9,
-                "unbearable": 9,
-                "complete": 10,
-                "total": 10,
-                "absolute": 10,
-            }
-
-            for term, rating in level_mapping.items():
-                if term in value_lower:
-                    return max(min_val, min(max_val, rating))
-
-        logger.warning(
-            "Could not normalize %s value '%s', using default %d",
-            field_name,
-            value,
-            default,
-        )
-        return default
-
-    async def create_from_voice_data(
-        self,
-        session_id: str,
-        user_id: str,
-        voice_data: RestDayDataExtraction | dict[str, Any],
-        transcript: str,
-        confidence_score: float,
-        processing_duration: float,
-    ) -> RestDayEntry:
-        """Create rest day entry from voice processing data."""
-        try:
-            # Handle both Pydantic model and dict input
-            if isinstance(voice_data, RestDayDataExtraction):
-                data = voice_data.model_dump()
-            else:
-                data = voice_data
-
-            # Normalize rest type
-            rest_type = self._normalize_rest_type(data.get("rest_type", "complete_rest"))
-
-            # Normalize integer fields that might come as strings from LLM
-            fatigue_level = self._normalize_integer_field(
-                data.get("fatigue_level"),
-                "fatigue_level",
-                1,
-                10,
-                5,
-            )
-            energy_level = self._normalize_integer_field(
-                data.get("energy_level"),
-                "energy_level",
-                1,
-                10,
-                5,
-            )
-            motivation_level = self._normalize_integer_field(
-                data.get("motivation_level"),
-                "motivation_level",
-                1,
-                10,
-                5,
-            )
-            soreness_level = (
-                self._normalize_integer_field(
-                    data.get("soreness_level"),
-                    "soreness_level",
-                    1,
-                    10,
-                    3,
-                )
-                if data.get("soreness_level") is not None
-                else None
-            )
-
-            # Create rest day entry
-            rest_entry = RestDayEntry(
-                session_id=session_id,
-                user_id=user_id,
-                # Voice processing metadata
-                transcript=transcript,
-                confidence_score=confidence_score,
-                processing_duration=processing_duration,
-                # Rest day data
-                rest_type=rest_type,
-                physical_state=data.get("physical_state", ""),
-                fatigue_level=fatigue_level,
-                energy_level=energy_level,
-                motivation_level=motivation_level,
-                mood_description=data.get("mood_description", ""),
-                mental_state=data.get("mental_state", "good"),
-                soreness_level=soreness_level,
-                training_reflections=data.get("training_reflections"),
-                goals_concerns=data.get("goals_concerns"),
-                recovery_activities=data.get("recovery_activities"),
-                notes=data.get("notes"),
-            )
-
-            self.session.add(rest_entry)
-            await self.session.commit()
-            await self.session.refresh(rest_entry)
-
-            logger.info("Created rest day entry from voice data: %s", rest_entry.id)
-            return rest_entry
-
-        except Exception as e:
-            await self.session.rollback()
-            logger.exception("Failed to create rest day entry from voice data")
-            msg = f"Failed to create rest day entry: {e}"
-            raise CricketRepositoryError(msg) from e
-
-    def _to_read_schema(self, db_record: RestDayEntry) -> RestDayEntryRead:
-        """Convert database model to read schema."""
-        return RestDayEntryRead(
-            id=db_record.id,
-            session_id=db_record.session_id,
-            user_id=db_record.user_id,
-            rest_type=db_record.rest_type.value,
-            physical_state=db_record.physical_state,
-            fatigue_level=db_record.fatigue_level,
-            energy_level=db_record.energy_level,
-            motivation_level=db_record.motivation_level,
-            mood_description=db_record.mood_description,
-            mental_state=db_record.mental_state,
-            soreness_level=db_record.soreness_level,
-            training_reflections=db_record.training_reflections,
-            goals_concerns=db_record.goals_concerns,
-            recovery_activities=db_record.recovery_activities,
-            notes=db_record.notes,
-            transcript=db_record.transcript,
-            confidence_score=db_record.confidence_score,
-            processing_duration=db_record.processing_duration,
-            created_at=db_record.created_at,
-            updated_at=db_record.updated_at,
-        )
-
-
-class CricketAnalyticsRepository:
-    """Repository for cricket-specific analytics operations."""
-
-    def __init__(self, session: AsyncSession) -> None:
-        """Initialize cricket analytics repository."""
-        self.session = session
-
-    async def get_cricket_analytics(
-        self,
-        user_id: str,
-        days_back: int = 30,
-    ) -> CricketAnalytics:
-        """Generate comprehensive cricket analytics."""
-        try:
-            cutoff_date = datetime.now(UTC) - timedelta(days=days_back)
-
-            # Get coaching session stats
-            coaching_result = await self.session.execute(
-                select(func.count(CricketCoachingEntry.id)).filter(
-                    CricketCoachingEntry.user_id == user_id,
-                    CricketCoachingEntry.created_at >= cutoff_date,
-                ),
-            )
-            total_coaching_sessions = coaching_result.scalar() or 0
-
-            # Get match stats
-            match_result = await self.session.execute(
-                select(func.count(CricketMatchEntry.id)).filter(
-                    CricketMatchEntry.user_id == user_id,
-                    CricketMatchEntry.created_at >= cutoff_date,
-                ),
-            )
-            total_matches = match_result.scalar() or 0
-
-            # Get rest day stats (import RestDayEntry if needed)
-            from fitness_tracking.models.cricket import RestDayEntry
-
-            rest_result = await self.session.execute(
-                select(func.count(RestDayEntry.id)).filter(
-                    RestDayEntry.user_id == user_id,
-                    RestDayEntry.created_at >= cutoff_date,
-                ),
-            )
-            total_rest_days = rest_result.scalar() or 0
-
-            # Calculate averages
-            avg_self_assessment = await self._calculate_average_self_assessment(
-                user_id,
-                cutoff_date,
-            )
-            batting_average = await self._calculate_batting_average(user_id, cutoff_date)
-            keeping_success_rate = await self._calculate_keeping_success_rate(user_id, cutoff_date)
-
-            # Get most practiced skill
-            most_practiced_skill = await self._get_most_practiced_skill(user_id, cutoff_date)
-
-            # Calculate confidence trend
-            confidence_trend = await self._calculate_confidence_trend(user_id, cutoff_date)
-
-            # Get improvement areas
-            improvement_areas = await self._get_improvement_areas(user_id, cutoff_date)
-
-            # Generate recommendations
-            recommendations = await self._generate_cricket_recommendations(user_id, cutoff_date)
-
-            return CricketAnalytics(
-                total_coaching_sessions=total_coaching_sessions,
-                total_matches=total_matches,
-                total_rest_days=total_rest_days,
-                average_self_assessment=round(avg_self_assessment, 1),
-                batting_average=batting_average,
-                keeping_success_rate=keeping_success_rate,
-                most_practiced_skill=most_practiced_skill,
-                confidence_trend=confidence_trend,
-                improvement_areas=improvement_areas,
-                recommendations=recommendations,
-            )
-
-        except Exception as e:
-            logger.exception("Failed to generate cricket analytics")
-            msg = f"Failed to generate cricket analytics: {e}"
-            raise CricketRepositoryError(msg) from e
-
-    async def _calculate_average_self_assessment(
-        self,
-        user_id: str,
-        cutoff_date: datetime,
-    ) -> float:
-        """Calculate average self assessment score."""
-        result = await self.session.execute(
-            select(func.avg(CricketCoachingEntry.self_assessment_score)).filter(
-                CricketCoachingEntry.user_id == user_id,
-                CricketCoachingEntry.created_at >= cutoff_date,
-                CricketCoachingEntry.self_assessment_score.is_not(None),
-            ),
-        )
-
-        avg_score = result.scalar()
-        return float(avg_score) if avg_score is not None else 5.0
-
-    async def _calculate_batting_average(
-        self,
-        user_id: str,
-        cutoff_date: datetime,
-    ) -> float | None:
-        """Calculate batting average from match data."""
-        result = await self.session.execute(
-            select(
-                func.sum(CricketMatchEntry.runs_scored),
-                func.count(CricketMatchEntry.id),
-            ).filter(
-                CricketMatchEntry.user_id == user_id,
-                CricketMatchEntry.created_at >= cutoff_date,
-                CricketMatchEntry.runs_scored.is_not(None),
-            ),
-        )
-
-        row = result.fetchone()
-        if row and row[0] is not None and row[1] > 0:
-            total_runs = row[0]
-            total_innings = row[1]
-            return round(total_runs / total_innings, 2)
-
-        return None
-
-    async def _calculate_keeping_success_rate(
-        self,
-        user_id: str,
-        cutoff_date: datetime,
-    ) -> float | None:
-        """Calculate wicket keeping success rate."""
-        result = await self.session.execute(
-            select(
-                func.sum(CricketMatchEntry.catches_taken),
-                func.sum(CricketMatchEntry.catches_dropped),
-            ).filter(
-                CricketMatchEntry.user_id == user_id,
-                CricketMatchEntry.created_at >= cutoff_date,
-                CricketMatchEntry.catches_taken.is_not(None),
-            ),
-        )
-
-        row = result.fetchone()
-        if row and row[0] is not None:
-            catches_taken = row[0] or 0
-            catches_dropped = row[1] or 0
-            total_chances = catches_taken + catches_dropped
-
-            if total_chances > 0:
-                return round((catches_taken / total_chances) * 100, 1)
-
-        return None
-
-    async def _get_most_practiced_skill(
-        self,
-        user_id: str,
-        cutoff_date: datetime,
-    ) -> str:
-        """Get the most practiced skill."""
-        result = await self.session.execute(
-            select(CricketCoachingEntry.skills_practiced).filter(
-                CricketCoachingEntry.user_id == user_id,
-                CricketCoachingEntry.created_at >= cutoff_date,
-            ),
-        )
-
-        skills_list = [row[0] for row in result.fetchall() if row[0]]
-        if not skills_list:
-            return "No skills recorded"
-
-        # Get most common skill (simple approach)
-        skill_counts = {}
-        for skill in skills_list:
-            # Split by common delimiters and count each skill
-            for individual_skill in skill.replace(",", " ").split():
-                individual_skill = individual_skill.strip().lower()
-                if individual_skill:
-                    skill_counts[individual_skill] = skill_counts.get(individual_skill, 0) + 1
-
-        if skill_counts:
-            most_practiced = max(skill_counts.items(), key=lambda x: x[1])[0]
-            return most_practiced.title()
-        return "General practice"
-
-    async def _calculate_confidence_trend(
-        self,
-        user_id: str,
-        cutoff_date: datetime,
-    ) -> dict[str, float]:
-        """Calculate confidence trend over time."""
-        result = await self.session.execute(
-            select(
-                CricketCoachingEntry.confidence_level,
-                CricketCoachingEntry.created_at,
-            )
-            .filter(
-                CricketCoachingEntry.user_id == user_id,
-                CricketCoachingEntry.created_at >= cutoff_date,
-            )
-            .order_by(CricketCoachingEntry.created_at),
-        )
-
-        confidence_data = [(row[0], row[1]) for row in result.fetchall() if row[0] is not None]
-
-        if len(confidence_data) < 2:
-            return {"trend": 0.0, "average": 5.0}
-
-        # Split into first and second half to calculate trend
-        mid_point = len(confidence_data) // 2
-        first_half_avg = sum(item[0] for item in confidence_data[:mid_point]) / mid_point
-        second_half_avg = sum(item[0] for item in confidence_data[mid_point:]) / (
-            len(confidence_data) - mid_point
-        )
-
-        trend = (
-            round(((second_half_avg - first_half_avg) / first_half_avg) * 100, 1)
-            if first_half_avg > 0
-            else 0.0
-        )
-        overall_avg = round(sum(item[0] for item in confidence_data) / len(confidence_data), 1)
-
-        return {"trend": trend, "average": overall_avg}
-
-    async def _get_improvement_areas(
-        self,
-        user_id: str,
-        cutoff_date: datetime,
-    ) -> list[str]:
-        """Get improvement areas from recent coaching sessions."""
-        result = await self.session.execute(
-            select(CricketCoachingEntry.areas_for_improvement)
-            .filter(
-                CricketCoachingEntry.user_id == user_id,
-                CricketCoachingEntry.created_at >= cutoff_date,
-            )
-            .limit(5),  # Get most recent 5 sessions
-        )
-
-        improvement_areas = []
-        for row in result.fetchall():
-            if row[0]:
-                # Split by common delimiters and extract individual areas
-                areas = [
-                    area.strip() for area in row[0].replace(",", ";").split(";") if area.strip()
-                ]
-                improvement_areas.extend(areas)
-
-        # Return unique areas (up to 5 most recent)
-        unique_areas = list(
-            dict.fromkeys(improvement_areas),
-        )  # Preserves order while removing duplicates
-        return unique_areas[:5] if unique_areas else ["No specific areas identified"]
-
-    async def _generate_cricket_recommendations(
-        self,
-        user_id: str,
-        cutoff_date: datetime,
-    ) -> list[str]:
-        """Generate cricket-specific recommendations."""
-        recommendations = []
-
-        try:
-            # Analyze confidence levels
-            confidence_result = await self.session.execute(
-                select(func.avg(CricketCoachingEntry.confidence_level)).filter(
-                    CricketCoachingEntry.user_id == user_id,
-                    CricketCoachingEntry.created_at >= cutoff_date,
-                ),
-            )
-
-            confidence = confidence_result.scalar()
-            if confidence and confidence < 6:
-                recommendations.append(
-                    "Focus on building confidence through regular practice and positive self-talk",
-                )
-
-            # Analyze frequency
-            sessions_result = await self.session.execute(
-                select(func.count(CricketCoachingEntry.id)).filter(
-                    CricketCoachingEntry.user_id == user_id,
-                    CricketCoachingEntry.created_at >= cutoff_date,
-                ),
-            )
-
-            sessions = sessions_result.scalar() or 0
-            days_period = (datetime.now(UTC) - cutoff_date).days
-            frequency = sessions / days_period if days_period > 0 else 0
-
-            if frequency < 0.2:  # Less than 2 sessions per 10 days
-                recommendations.append(
-                    "Try to increase your cricket practice frequency for better skill development",
-                )
-
-            if not recommendations:
-                recommendations.append(
-                    "Keep up the excellent cricket training! Your consistency is paying off",
-                )
-
-        except Exception:
-            logger.exception("Error generating cricket recommendations")
-            recommendations.append("Unable to generate cricket recommendations at this time")
-
-        return recommendations
