+++ fitness_tracking/schemas/enums/coaching_focus_type.py
@@ -1,0 +1,12 @@
+from enum import Enum
+
+
+class CoachingFocus(str, Enum):
+    """Focus areas for coaching sessions."""
+
+    TECHNIQUE = "technique"
+    FITNESS = "fitness"
+    TACTICS = "tactics"
+    MENTAL = "mental"
+    SPECIFIC_SKILLS = "specific_skills"
+    GAME_SITUATION = "game_situation"
