+++ fitness_tracking/schemas/enums/exercise_type.py
