+++ fitness_tracking/schemas/enums/intensity_level.py
