+++ fitness_tracking/schemas/rest_day.py
@@ -1,0 +1,108 @@
+from typing import Any
+
+from pydantic import Field
+
+from common.schemas import AppBaseModel, TimestampBase
+from common.schemas.activity import ActivityEntryBase
+from common.schemas.entry_type import EntryType
+from common.schemas.primary_key_base import PrimaryKeyBase
+
+
+class RestDayEntryBase(ActivityEntryBase):
+    """Base schema for rest day entries."""
+
+    entry_type: EntryType = Field(default=EntryType.REST_DAY, description="Entry type")
+    rest_type: str = Field(..., description="Type of rest (active, complete, partial)")
+    planned: bool = Field(default=False, description="Whether this was a planned rest day")
+
+
+class RestDayEntryCreate(RestDayEntryBase):
+    """Schema for creating a rest day entry."""
+
+    # Recovery activities
+    recovery_activities: list[str] | None = Field(None, description="Recovery activities performed")
+    sleep_hours: float | None = Field(None, ge=0.0, le=24.0, description="Hours of sleep")
+    sleep_quality: int | None = Field(None, ge=1, le=10, description="Sleep quality 1-10")
+
+    # Physical state
+    muscle_soreness: int | None = Field(None, ge=1, le=10, description="Muscle soreness level 1-10")
+    fatigue_level: int | None = Field(None, ge=1, le=10, description="Fatigue level 1-10")
+    stress_level: int | None = Field(None, ge=1, le=10, description="Stress level 1-10")
+
+    # Recovery metrics
+    recovery_score: int | None = Field(None, ge=1, le=100, description="Recovery score 1-100")
+    readiness_for_next_workout: int | None = Field(
+        None,
+        ge=1,
+        le=10,
+        description="Readiness for next workout 1-10",
+    )
+
+    # Wellness activities
+    meditation_minutes: int | None = Field(None, ge=0, description="Meditation time in minutes")
+    stretching_minutes: int | None = Field(None, ge=0, description="Stretching time in minutes")
+    massage_minutes: int | None = Field(None, ge=0, description="Massage time in minutes")
+
+    # Nutrition focus
+    hydration_liters: float | None = Field(None, ge=0.0, description="Water intake in liters")
+    protein_focus: bool | None = Field(None, description="Whether focused on protein intake")
+    nutrition_notes: str | None = Field(None, description="Nutrition notes")
+
+    # Data quality tracking
+    processing_duration: float | None = Field(None, ge=0.0, description="Processing duration")
+    data_quality_score: float | None = Field(None, ge=0.0, le=1.0, description="Data quality score")
+    manual_overrides: dict[str, Any] | None = Field(None, description="Manual data overrides")
+    validation_notes: str | None = Field(None, description="Validation notes")
+    energy_level: int | None = Field(None, ge=1, le=10, description="Energy level 1-10")
+    notes: str | None = Field(None, description="Additional notes")
+
+
+class RestDayEntryUpdate(AppBaseModel):
+    """Schema for updating a rest day entry."""
+
+    rest_type: str | None = None
+    planned: bool | None = None
+    recovery_activities: list[str] | None = None
+    sleep_hours: float | None = Field(None, ge=0.0, le=24.0)
+    sleep_quality: int | None = Field(None, ge=1, le=10)
+    muscle_soreness: int | None = Field(None, ge=1, le=10)
+    fatigue_level: int | None = Field(None, ge=1, le=10)
+    stress_level: int | None = Field(None, ge=1, le=10)
+    recovery_score: int | None = Field(None, ge=1, le=100)
+    readiness_for_next_workout: int | None = Field(None, ge=1, le=10)
+    meditation_minutes: int | None = Field(None, ge=0)
+    stretching_minutes: int | None = Field(None, ge=0)
+    massage_minutes: int | None = Field(None, ge=0)
+    hydration_liters: float | None = Field(None, ge=0.0)
+    protein_focus: bool | None = None
+    nutrition_notes: str | None = None
+    data_quality_score: float | None = Field(None, ge=0.0, le=1.0)
+    manual_overrides: dict[str, Any] | None = None
+    validation_notes: str | None = None
+    energy_level: int | None = Field(None, ge=1, le=10)
+    notes: str | None = None
+
+
+class RestDayEntryRead(PrimaryKeyBase, TimestampBase, RestDayEntryBase):
+    """Schema for reading rest day entry data."""
+
+    recovery_activities: list[str] | None
+    sleep_hours: float | None
+    sleep_quality: int | None
+    muscle_soreness: int | None
+    fatigue_level: int | None
+    stress_level: int | None
+    recovery_score: int | None
+    readiness_for_next_workout: int | None
+    meditation_minutes: int | None
+    stretching_minutes: int | None
+    massage_minutes: int | None
+    hydration_liters: float | None
+    protein_focus: bool | None
+    nutrition_notes: str | None
+    processing_duration: float | None
+    data_quality_score: float | None
+    manual_overrides: dict[str, Any] | None
+    validation_notes: str | None
+    energy_level: int | None
+    notes: str | None
