+++ voice_processing/services/conversation_service.py
@@ -10,12 +10,12 @@ from openai.types.chat.parsed_chat_completion import ParsedChatCompletion
 
 from common.config.settings import settings
 from common.exceptions import AppError
-from fitness_tracking.schemas.cricket import (
+from fitness_tracking.schemas.cricket_coaching import (
     CricketCoachingDataExtraction,
-    CricketMatchDataExtraction,
-    RestDayDataExtraction,
 )
-from fitness_tracking.schemas.fitness import FitnessDataExtraction
+from fitness_tracking.schemas.cricket_match_data_extraction import CricketMatchDataExtraction
+from fitness_tracking.schemas.fitness_data_extraction import FitnessDataExtraction
+from fitness_tracking.schemas.rest_day_data_extraction import RestDayDataExtraction
 from voice_processing.schemas.conversation import (
     ActivityType,
     ConversationAnalysis,
